Smaller images

+++ test.sh
@@ -24,7 +24,7 @@ TOS=($(jq -r .[].to steps.json))
 COUNT=${#LANGUAGES[@]}
 
 cd test
-for i in $(seq 0 1 $(($COUNT-1))); do
+for i in $(seq 94 1 $(($COUNT-1))); do
 	LANGUAGE="${LANGUAGES[$i]}"
 	CONTAINER="${CONTAINERS[$i]}"
 	IMAGE="${IMAGES[$i]}"
@@ -33,6 +33,10 @@ for i in $(seq 0 1 $(($COUNT-1))); do
 	TO="${TOS[$i]}"
 
 	echo "Run $i $LANGUAGE..."
+	if [[ "$IMAGE" == "quine/all" ]]; then
+		echo $COMMAND
+		exit
+	fi
 	docker run --rm -ti -v $(pwd)/$FROM:/root/go/src/github.com/dgageot/quine-relay/$FROM:ro -v $(pwd):/output $IMAGE bash -c "$COMMAND && cp $TO /output/$TO"
 done
